refactor: rename `/config/client.ts` to `/config/common.ts`
