+++ config/features.ts
@@ -1,4 +1,4 @@
-import { commonConfig } from "@/config/client"
+import { commonConfig } from "@/config/common"
 
 // Basic feature flags for now
 // But likely to stay as a separate global variable like this
